AI_service/requirements.txt
tensorflow==2.15.0
keras==2.15.0
opencv-python
Pillow
numpy
scipy
pandas
scikit-learn
matplotlib
seaborn
tqdm
protobuf==4.25.0
h5py
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart
